[no ci]  Keep removing query manager generic
