+++ full-node/db/sov-db/src/schema/mod.rs
@@ -4,4 +4,6 @@ pub mod tables;
 /// are split into their constituent parts and stored in separate tables for easy retrieval.
 pub mod types;
 
-pub use sov_schema_db::snapshot::{CacheDb, ChangeSet, QueryManager};
+pub use sov_schema_db::cache::cache_container::CacheContainer;
+pub use sov_schema_db::cache::cache_db::CacheDb;
+pub use sov_schema_db::cache::change_set::ChangeSet;
