+++ full-node/db/sov-db/src/state_db.rs
@@ -3,7 +3,8 @@ use std::sync::{Arc, Mutex};
 
 use jmt::storage::{HasPreimage, TreeReader, TreeWriter};
 use jmt::{KeyHash, Version};
-use sov_schema_db::snapshot::{CacheDb, ChangeSet, QueryManager};
+use sov_schema_db::cache::cache_db::CacheDb;
+use sov_schema_db::cache::change_set::ChangeSet;
 use sov_schema_db::SchemaBatch;
 
 use crate::rocks_db_config::gen_rocksdb_options;
@@ -16,9 +17,9 @@ use crate::schema::types::StateKey;
 /// StateDB implements several convenience functions for state storage -
 /// notably the [`TreeReader`] and [`TreeWriter`] traits.
 #[derive(Debug)]
-pub struct StateDB<Q> {
+pub struct StateDB {
     /// The underlying [`CacheDb`] that plays as local cache and pointer to previous snapshots and/or [`sov_schema_db::DB`]
-    db: Arc<CacheDb<Q>>,
+    db: Arc<CacheDb>,
     /// The [`Version`] that will be used for the next batch of writes to the DB
     /// This [`Version`] is also used for querying data,
     /// so if this instance of StateDB is used as read only, it won't see newer data.
@@ -26,7 +27,7 @@ pub struct StateDB<Q> {
 }
 
 // Manual implementation of [`Clone`] to satisfy compiler
-impl<Q> Clone for StateDB<Q> {
+impl Clone for StateDB {
     fn clone(&self) -> Self {
         StateDB {
             db: self.db.clone(),
@@ -35,7 +36,7 @@ impl<Q> Clone for StateDB<Q> {
     }
 }
 
-impl<Q> StateDB<Q> {
+impl StateDB {
     const DB_PATH_SUFFIX: &'static str = "state";
     const DB_NAME: &'static str = "state-db";
 
@@ -53,15 +54,13 @@ impl<Q> StateDB<Q> {
     /// Convert it to [`ChangeSet`] which cannot be edited anymore
     pub fn freeze(self) -> anyhow::Result<ChangeSet> {
         let inner = Arc::into_inner(self.db).ok_or(anyhow::anyhow!(
-            "StateDB underlying DbSnapshot has more than 1 strong references"
+            "StateDB underlying CacheDb has more than 1 strong references"
         ))?;
         Ok(ChangeSet::from(inner))
     }
-}
 
-impl<Q: QueryManager> StateDB<Q> {
     /// Creating instance of [`StateDB`] from [`CacheDb`]
-    pub fn with_db_snapshot(db_snapshot: CacheDb<Q>) -> anyhow::Result<Self> {
+    pub fn with_db_snapshot(db_snapshot: CacheDb) -> anyhow::Result<Self> {
         let next_version = Self::next_version_from(&db_snapshot)?;
         Ok(Self {
             db: Arc::new(db_snapshot),
@@ -115,7 +114,7 @@ impl<Q: QueryManager> StateDB<Q> {
         *version
     }
 
-    fn next_version_from(db_snapshot: &CacheDb<Q>) -> anyhow::Result<Version> {
+    fn next_version_from(db_snapshot: &CacheDb) -> anyhow::Result<Version> {
         let last_key_value = db_snapshot.get_largest::<JmtNodes>()?;
         let largest_version = last_key_value.map(|(k, _)| k.version());
         let next_version = largest_version
@@ -126,7 +125,7 @@ impl<Q: QueryManager> StateDB<Q> {
     }
 }
 
-impl<Q: QueryManager> TreeReader for StateDB<Q> {
+impl TreeReader for StateDB {
     fn get_node_option(
         &self,
         node_key: &jmt::storage::NodeKey,
@@ -153,7 +152,7 @@ impl<Q: QueryManager> TreeReader for StateDB<Q> {
     }
 }
 
-impl<Q: QueryManager> TreeWriter for StateDB<Q> {
+impl TreeWriter for StateDB {
     fn write_node_batch(&self, node_batch: &jmt::storage::NodeBatch) -> anyhow::Result<()> {
         let mut batch = SchemaBatch::new();
         for (node_key, node) in node_batch.nodes() {
@@ -174,7 +173,7 @@ impl<Q: QueryManager> TreeWriter for StateDB<Q> {
     }
 }
 
-impl<Q: QueryManager> HasPreimage for StateDB<Q> {
+impl HasPreimage for StateDB {
     fn preimage(&self, key_hash: KeyHash) -> anyhow::Result<Option<Vec<u8>>> {
         self.db.read::<KeyHashToKey>(&key_hash.0)
     }
@@ -186,7 +185,7 @@ mod state_db_tests {
 
     use jmt::storage::{NodeBatch, TreeReader, TreeWriter};
     use jmt::KeyHash;
-    use sov_schema_db::snapshot::{CacheDb, NoopQueryManager, ReadOnlyLock};
+    use sov_schema_db::cache::{CacheDb, NoopQueryManager, ReadOnlyLock};
 
     use super::StateDB;
